# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Yercaud Business Directory API
  version: 0.1.0
  description: >
    Backend contract for the public frontend and Admin panel. Auth is
    httpOnly session cookies (ADR: session cookies over JWT). See
    /CONTEXT.md and /docs/adr/ for the domain model and design decisions
    this contract implements.
servers:
  - url: http://localhost:4000/v1
    description: Local dev
security:
  - cookieAuth: []
...
paths:
...
  /locations:
    get:
      tags: [Taxonomies]
      security: []
      summary: List locations (flat, Yercaud-only — CONTEXT.md)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NamedLookup' }
    post:
      tags: [Taxonomies]
      summary: Create a location (Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NamedLookupInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NamedLookup' }
  /locations/{locationId}:
    parameters:
      - name: locationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Taxonomies]
      summary: Update a location (Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NamedLookupInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NamedLookup' }
    delete:
      tags: [Taxonomies]
      summary: Delete a location (Super Admin); Listings referencing it have location_id set to null (ADR 0009)
      responses:
        '204': { description: Deleted }
...
  /price-bands:
    get:
      tags: [Taxonomies]
      security: []
      summary: List price bands
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PriceBand' }
    post:
      tags: [Taxonomies]
      summary: Create a price band (Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PriceBand' }
      responses:
        '201': { description: Created }
  /price-bands/{priceBandId}:
    parameters:
      - name: priceBandId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Taxonomies]
      summary: Update a price band (Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PriceBand' }
      responses:
        '200': { description: OK }
    delete:
      tags: [Taxonomies]
      summary: Delete a price band (Super Admin)
      responses:
        '204': { description: Deleted }

  # ===========================================================================
  # Businesses
  # ===========================================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
...
  schemas:
...
    NamedLookup:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        # Present on Categories and Locations; absent on the plain lookups
        # (amenities, attribute tags, property types, languages), which have no
        # slug column.
        slug: { type: string }
    NamedLookupInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
    PriceBand:
      type: object
      required: [id, label, minAmount, maxAmount]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        minAmount: { type: integer }
        maxAmount: { type: integer, nullable: true }
...